+++ packages/components/src/components/action/button-pure/button-pure-utils.ts
@@ -1,7 +1,6 @@
 import type { ButtonAriaAttributes } from '../../../utils';
 import type { AriaAttributes, LinkButtonPureIconName, SelectedAriaAttributes } from '../../../types';
 import { getButtonBaseAriaAttributes, getTagName, hasVisibleIcon, parseAndGetAriaAttributes } from '../../../utils';
-import { BUTTON_ARIA_ATTRIBUTES } from '../../../utils';
 
 export const warnIfIsLoadingAndIconIsNone = (
   host: HTMLElement,
@@ -24,7 +23,7 @@ export const getButtonAriaAttributes = (
   aria: SelectedAriaAttributes<ButtonAriaAttributes>
 ): AriaAttributes => {
   return {
-    ...parseAndGetAriaAttributes(aria, BUTTON_ARIA_ATTRIBUTES),
+    ...parseAndGetAriaAttributes(aria),
     ...getButtonBaseAriaAttributes(isDisabled, isLoading),
     'aria-describedby': hasSubline ? 'subline' : null,
   };