+++ packages/components/src/components/action/button/button-utils.ts
@@ -1,7 +1,7 @@
 import type { SelectedAriaAttributes } from '../../../types';
 import type { AriaAttributes } from '../../../aria-types';
 import type { ButtonAriaAttributes } from '../../../utils';
-import { BUTTON_ARIA_ATTRIBUTES, getButtonBaseAriaAttributes, parseAndGetAriaAttributes } from '../../../utils';
+import { getButtonBaseAriaAttributes, parseAndGetAriaAttributes } from '../../../utils';
 
 export const getButtonAriaAttributes = (
   isDisabled: boolean,
@@ -9,7 +9,7 @@ export const getButtonAriaAttributes = (
   aria: SelectedAriaAttributes<ButtonAriaAttributes>
 ): AriaAttributes => {
   return {
-    ...parseAndGetAriaAttributes(aria, BUTTON_ARIA_ATTRIBUTES),
+    ...parseAndGetAriaAttributes(aria),
     ...getButtonBaseAriaAttributes(isDisabled, isLoading),
   };
 };