+++ packages/components/src/components/feedback/spinner/spinner-utils.ts
@@ -1,9 +1,8 @@
 import type { BreakpointKey, BreakpointCustomizable } from '../../../types';
 import { BREAKPOINTS, parseJSON } from '../../../utils';
 
-const SPINNER_SIZES = ['small', 'medium', 'large', 'inherit'] as const;
+export const SPINNER_SIZES = ['small', 'medium', 'large', 'inherit'] as const;
 export type SpinnerSizeType = typeof SPINNER_SIZES[number];
-
 export type SpinnerSize = BreakpointCustomizable<SpinnerSizeType>;
 
 export const verifySpinnerSize = (spinnerSize: SpinnerSize): void => {