+++ packages/components/CHANGELOG.md
@@ -9,8 +9,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### [Unreleased]
 
+#### Changed
+
 - Font loading strategy for Porsche Next uses `font-display: swap` (FOUT) instead of `font-display: block` for better
   performance and less potential for layout shifts
+
+#### Added
+
+- Validation of properties for all components
+
 
 ### [2.14.0-beta.0] - 2022-07-12